Split PackageController into four files
