+++ Sources/App/Controllers/PackageController+ShowRoute.swift
@@ -18,7 +18,6 @@ import Vapor
 
 
 extension PackageController {
-    typealias PackageResult = Joined5<Package, Repository, DefaultVersion, ReleaseVersion, PreReleaseVersion>
 
     enum ShowRoute {
 
@@ -169,6 +168,7 @@ private extension Array where Element == PackageController.BuildsRoute.BuildInfo
     typealias SwiftVersionResults = PackageShow.Model.SwiftVersionResults
     typealias PlatformResults = PackageShow.Model.PlatformResults
 
+#warning("add test")
     func platformBuildResults(kind: Version.Kind) -> NamedBuildResults<PlatformResults>? {
         let builds = filter { $0.versionKind == kind}
         // builds of the same kind all originate from the same Version via a join,
@@ -195,6 +195,7 @@ private extension Array where Element == PackageController.BuildsRoute.BuildInfo
             )
     }
 
+#warning("add test")
     func swiftVersionBuildResults(kind: Version.Kind) -> NamedBuildResults<SwiftVersionResults>? {
         let builds = filter { $0.versionKind == kind}
         // builds of the same kind all originate from the same Version via a join,
@@ -222,25 +223,6 @@ private extension Array where Element == PackageController.BuildsRoute.BuildInfo
 }
 
 
-extension Array where Element == PackageController.BuildsRoute.BuildInfo {
-    var noneSucceeded: Bool {
-        allSatisfy { $0.status != .ok }
-    }
-
-    var anySucceeded: Bool {
-        !noneSucceeded
-    }
-
-    var nonePending: Bool {
-        allSatisfy { $0.status.isCompleted }
-    }
-
-    var anyPending: Bool {
-        !nonePending
-    }
-}
-
-
 private extension Build.Platform {
     func isCompatible(with other: PackageShow.Model.PlatformCompatibility) -> Bool {
         switch self {
@@ -261,145 +243,20 @@ private extension Build.Platform {
 }
 
 
-extension PackageController.PackageResult {
-    var package: Package { model }
-    // We can force-unwrap due to the inner join
-    var repository: Repository { relation1! }
-    // We can force-unwrap due to the inner join
-    var defaultBranchVersion: DefaultVersion { relation2! }
-    var releaseVersion: ReleaseVersion? { relation3 }
-    var preReleaseVersion: PreReleaseVersion? { relation4 }
-
-    @available(*, deprecated)
-    var versions: [Version] {
-        [defaultBranchVersion.model, releaseVersion?.model, preReleaseVersion?.model].compactMap { $0 }
+private extension Array where Element == PackageController.BuildsRoute.BuildInfo {
+    var noneSucceeded: Bool {
+        allSatisfy { $0.status != .ok }
     }
 
-    static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<Self> {
-        Package.query(on: database)
-            .join(Repository.self,
-                  on: \Repository.$package.$id == \Package.$id,
-                  method: .inner)
-            .join(DefaultVersion.self,
-                  on: \DefaultVersion.$package.$id == \Package.$id,
-                  method: .inner)
-            .join(Package.self, ReleaseVersion.self,
-                  on: .custom(#"""
-                    LEFT JOIN "\#(Version.schema)" AS "\#(ReleaseVersion.name)"
-                    ON "\#(Package.schema)"."id" = "\#(ReleaseVersion.name)"."package_id"
-                    AND "\#(ReleaseVersion.name)"."latest" = 'release'
-                    """#)
-            )
-            .join(Package.self, PreReleaseVersion.self,
-                  on: .custom(#"""
-                    LEFT JOIN "\#(Version.schema)" AS "\#(PreReleaseVersion.name)"
-                    ON "\#(Package.schema)"."id" = "\#(PreReleaseVersion.name)"."package_id"
-                    AND "\#(PreReleaseVersion.name)"."latest" = 'pre_release'
-                    """#)
-            )
-            .filter(Repository.self, \.$owner, .custom("ilike"), owner)
-            .filter(Repository.self, \.$name, .custom("ilike"), repository)
-            .filter(DefaultVersion.self, \.$latest == .defaultBranch)
-        // TODO: only load required fields
-            .first()
-            .unwrap(or: Abort(.notFound))
-            .map(Self.init(model:))
+    var anySucceeded: Bool {
+        !noneSucceeded
     }
-}
 
+    var nonePending: Bool {
+        allSatisfy { $0.status.isCompleted }
+    }
 
-final class DefaultVersion: ModelAlias {
-    static let name = "default_version"
-    let model = Version()
-
-    #warning("temp. to make it compile")
-    var products: [Product] { [] }
-}
-
-final class ReleaseVersion: ModelAlias {
-    static let name = "release_version"
-    let model = Version()
-}
-
-final class PreReleaseVersion: ModelAlias {
-    static let name = "pre_release_version"
-    let model = Version()
-}
-
-
-extension PackageController {
-    enum BuildsRoute {
-        struct PackageInfo: Equatable {
-            var packageName: String?
-            var repositoryOwner: String
-            var repositoryName: String
-
-            static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<(PackageInfo)> {
-                Joined3<Package, Repository, Version>
-                    .query(on: database, owner: owner, repository: repository, version: .defaultBranch)
-                    .field(Repository.self, \.$owner)
-                    .field(Repository.self, \.$name)
-                    .field(Version.self, \.$packageName)
-                    .first()
-                    .unwrap(or: Abort(.notFound))
-                    .flatMapThrowing { model in
-                        let repo = model.repository
-                        guard let repoOwner = repo.owner,
-                              let repoName = repo.name else {
-                                  throw Abort(.notFound)
-                              }
-                        return .init(packageName: model.version.packageName,
-                                     repositoryOwner: repoOwner,
-                                     repositoryName: repoName)
-                    }
-            }
-        }
-
-        struct BuildInfo: Equatable {
-            var versionKind: Version.Kind
-            var reference: Reference
-            var buildId: Build.Id
-            var swiftVersion: SwiftVersion
-            var platform: Build.Platform
-            var status: Build.Status
-
-            static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<[BuildInfo]> {
-                Joined4<Build, Version, Package, Repository>
-                    .query(on: database)
-                    .filter(Version.self, \Version.$latest != nil)
-                    .filter(Repository.self, \.$owner, .custom("ilike"), owner)
-                    .filter(Repository.self, \.$name, .custom("ilike"), repository)
-                    .field(\.$id)
-                    .field(\.$swiftVersion)
-                    .field(\.$platform)
-                    .field(\.$status)
-                    .field(Version.self, \.$latest)
-                    .field(Version.self, \.$packageName)
-                    .field(Version.self, \.$reference)
-                    .all()
-                    .flatMapThrowing { results in
-                        try results
-                            .compactMap { res -> BuildInfo? in
-                                let build = res.build
-                                let version = res.version
-                                guard let kind = version.latest,
-                                      let reference = version.reference else {
-                                          return nil
-                                      }
-                                return try BuildInfo(versionKind: kind,
-                                                     reference: reference,
-                                                     buildId: build.requireID(),
-                                                     swiftVersion: build.swiftVersion,
-                                                     platform: build.platform,
-                                                     status: build.status)
-                            }
-                    }
-            }
-        }
-
-        static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<(PackageInfo, [BuildInfo])> {
-            PackageInfo.query(on: database, owner: owner, repository: repository)
-                .and(BuildInfo.query(on: database, owner: owner, repository: repository))
-        }
+    var anyPending: Bool {
+        !nonePending
     }
 }
